+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.0.8
+## 0.0.9
 
 * Added support for both ListView and GridView implementations
 * Implemented comprehensive error handling for network operations
@@ -10,3 +10,6 @@
 * handed network status with stream
 * change controller and main class updating structure
 * fully documented structured
+* enhance error handler and onError function which can make full control in coming error by give ApiRequestException
+* remove unwanted toast and make it optional based developer needs
+* give BuildContext in important callBacks